requirements.txt
simpy==4.0.1
PyYAML==6.0
beautifultable==1.0.1